+++ src/main/java/org/smoothbuild/cli/StandardOptions.java
@@ -1,14 +1,26 @@
 package org.smoothbuild.cli;
 
-import picocli.CommandLine.Command;
+import org.smoothbuild.cli.console.Level;
+import org.smoothbuild.cli.console.LogLevelConverter;
 
-@Command(
-    usageHelpAutoWidth= true,
-    synopsisHeading = "@|bold,underline Usage:|@%n",
-    descriptionHeading = "%n@|bold,underline Description:|@%n",
-    commandListHeading = "%n@|bold,underline Commands:|@%n",
-    parameterListHeading = "%n@|bold,underline Parameters:|@%n",
-    optionListHeading = "%n@|bold,underline Options:|@%n"
-)
-public class StandardOptions {
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
+
+@Command
+public class StandardOptions extends FormattedHeadings {
+  @Option(
+      names = { "--log-level", "-l" },
+      defaultValue = "info",
+      paramLabel = "<level>",
+      converter = LogLevelConverter.class,
+      description =
+          "Show logs with specified level or above.\n" +
+          "\n" +
+          "Available levels:\n" +
+          "  f, fatal   - show FATAL logs\n" +
+          "  e, error   - show FATAL, ERROR logs\n" +
+          "  w, warning - show FATAL, ERROR, WARNING logs\n" +
+          "  i, info    - show FATAL, ERROR, WARNING, INFO logs\n"
+  )
+  Level logLevel;
 }