+++ src/main/java/org/smoothbuild/exec/task/base/NonComputableTask.java
@@ -5,11 +5,11 @@ import java.util.List;
 import org.smoothbuild.lang.base.Location;
 import org.smoothbuild.lang.object.type.ConcreteType;
 
-public abstract class NonComputableTask extends BuildTask {
+public abstract class NonComputableTask extends Task {
   private final String name;
   private final ConcreteType type;
 
-  public NonComputableTask(String name, ConcreteType type, List<? extends BuildTask> dependencies,
+  public NonComputableTask(String name, ConcreteType type, List<? extends Task> dependencies,
       Location location) {
     super(dependencies, location);
     this.name = name;