+++ src/main/java/org/smoothbuild/exec/task/base/TaskKind.java
@@ -1,11 +1,7 @@
 package org.smoothbuild.exec.task.base;
 
 public enum TaskKind {
-  VALIDATING,
-  PARSING,
-  BUILDING_LITERAL,
-  BUILDING_CONVERSION,
-  BUILDING_NATIVE_CALL,
-  BUILDING_NORMAL_CALL,
-  SAVING
+  CALL,
+  CONVERSION,
+  LITERAL,
 }