+++ src/main/java/org/smoothbuild/exec/task/parallel/ExecutionReporter.java
@@ -4,7 +4,6 @@ import static com.google.common.base.Throwables.getStackTraceAsString;
 import static java.util.stream.Collectors.toList;
 import static org.smoothbuild.cli.console.Log.error;
 import static org.smoothbuild.cli.console.Log.fatal;
-import static org.smoothbuild.lang.object.base.Messages.isEmpty;
 import static org.smoothbuild.lang.object.base.Messages.level;
 import static org.smoothbuild.lang.object.base.Messages.text;
 
@@ -16,8 +15,8 @@ import org.smoothbuild.cli.console.Console;
 import org.smoothbuild.cli.console.Log;
 import org.smoothbuild.cli.console.Reporter;
 import org.smoothbuild.exec.comp.MaybeOutput;
-import org.smoothbuild.exec.task.base.BuildTask;
 import org.smoothbuild.exec.task.base.Computed;
+import org.smoothbuild.exec.task.base.Task;
 import org.smoothbuild.lang.object.base.Array;
 import org.smoothbuild.lang.object.base.Struct;
 
@@ -36,40 +35,41 @@ public class ExecutionReporter {
     this.reporter = reporter;
   }
 
-  public void report(BuildTask task, Computed computed) {
+  public void report(Task task, Computed computed) {
     MaybeOutput maybeOutput = computed.computed();
     boolean fromCache = computed.isFromCache();
     if (maybeOutput.hasOutput()) {
-      Array messages = maybeOutput.output().messages();
-      if (!isEmpty(messages)) {
-        print(task, fromCache, messages);
-      }
+      print(task, fromCache, maybeOutput.output().messages());
     } else {
       Log error = error(
           "Execution failed with:\n" + getStackTraceAsString(maybeOutput.exception()));
-      reporter.report(header(task, fromCache), List.of(error));
+      print(task, fromCache, List.of(error));
     }
   }
 
-  public void reportComputerException(BuildTask task, Throwable throwable) {
+  public void reportComputerException(Task task, Throwable throwable) {
     Log fatal = fatal(
         "Internal smooth error, computation failed with:" + getStackTraceAsString(throwable));
-    reporter.report(header(task, ""), List.of(fatal));
+    reporter.report(task, header(task, ""), List.of(fatal));
   }
 
-  private void print(BuildTask task, boolean fromCache, Array messages) {
+  private void print(Task task, boolean fromCache, Array messages) {
     List<Log> logs = Streams.stream(messages.asIterable(Struct.class))
         .map(m -> new Log(level(m), text(m)))
         .collect(toList());
-    reporter.report(header(task, fromCache), logs);
+    print(task, fromCache, logs);
+  }
+
+  public void print(Task task, boolean fromCache, List<Log> logs) {
+    reporter.report(task, header(task, fromCache), logs);
   }
 
   @VisibleForTesting
-  static String header(BuildTask task, boolean fromCache) {
+  static String header(Task task, boolean fromCache) {
     return header(task, fromCache ? " CACHED" : "");
   }
 
-  private static String header(BuildTask task, String cacheStatus) {
+  private static String header(Task task, String cacheStatus) {
     String locationString = task.location().toString();
     int paddedLength = Console.MESSAGE_GROUP_NAME_HEADER_LENGTH - locationString.length();
     String name = Strings.padEnd(task.name(), paddedLength, ' ');