+++ src/main/java/org/smoothbuild/parse/expr/AccessorCallExpression.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.smoothbuild.exec.comp.AccessorCallAlgorithm;
 import org.smoothbuild.exec.comp.Algorithm;
-import org.smoothbuild.exec.task.base.BuildTask;
 import org.smoothbuild.exec.task.base.NormalTask;
+import org.smoothbuild.exec.task.base.Task;
 import org.smoothbuild.lang.base.Accessor;
 import org.smoothbuild.lang.base.Location;
 import org.smoothbuild.lang.base.Scope;
@@ -20,9 +20,9 @@ public class AccessorCallExpression extends Expression {
   }
 
   @Override
-  public BuildTask createTask(Scope<BuildTask> scope) {
+  public Task createTask(Scope<Task> scope) {
     Algorithm algorithm = new AccessorCallAlgorithm(accessor);
-    List<BuildTask> dependencies = childrenTasks(scope);
+    List<Task> dependencies = childrenTasks(scope);
     return new NormalTask(algorithm, dependencies, location(), true);
   }
 }