+++ src/main/java/org/smoothbuild/parse/expr/LiteralExpression.java
@@ -2,8 +2,8 @@ package org.smoothbuild.parse.expr;
 
 import static org.smoothbuild.util.Strings.escapedAndLimitedWithEllipsis;
 
-import org.smoothbuild.exec.task.base.BuildTask;
 import org.smoothbuild.exec.task.base.FixedTask;
+import org.smoothbuild.exec.task.base.Task;
 import org.smoothbuild.lang.base.Location;
 import org.smoothbuild.lang.base.Scope;
 import org.smoothbuild.lang.object.base.SObject;
@@ -18,7 +18,7 @@ public class LiteralExpression extends Expression {
   }
 
   @Override
-  public BuildTask createTask(Scope<BuildTask> scope) {
+  public Task createTask(Scope<Task> scope) {
     String name = escapedAndLimitedWithEllipsis(((SString) this.object).jValue(), 20);
     return new FixedTask(object, name, location());
   }